toolkit/content/tests/widgets/test_videocontrols_audio.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Video controls with Audio file test</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>

<div id="content">
  <video id="video" controls preload="metadata"></video>
</div>

<pre id="test">
<script class="testbody" type="application/javascript">

  const video = document.getElementById("video");

  function findElementByAttribute(element, aName, aValue) {
    if (!("getAttribute" in element)) {
      return false;
    }
    if (element.getAttribute(aName) === aValue) {
      return element;
    }
    let children =
      InspectorUtils.getChildrenForNode(element, true);
    for (let child of children) {
      var result = findElementByAttribute(child, aName, aValue);
      if (result) {
        return result;
      }
    }
    return false;
  }

  function loadedmetadata(event) {
    SimpleTest.executeSoon(async function test_fullscreen_unavailable() {
      const { x, y } = video.getBoundingClientRect();
      const controlBar = findElementByAttribute(video, "class", "controlBar");
      is(controlBar.getAttribute("fullscreen-unavailable"), "true", "Fullscreen button is hidden");
      const eventLogger = new EventLogger(2);
      video.addEventListener("play", eventLogger);
      video.addEventListener("pause", eventLogger);
      document.addEventListener("mozfullscreenchange", eventLogger);
      synthesizeMouse(video, x, y, { clickCount: 2 });
      const events = await eventLogger.waitForExpectedEvents();
      ok(
        !events.find(e => e.type === "mozfullscreenchange"),
        "Double clicking should not trigger fullscreen event"
      );
      is(events.length, 2, "Two events should have been fired.");
      SimpleTest.finish();
    });
  }

  SpecialPowers.pushPrefEnv({"set": [["media.cache_size", 40000]]}, startTest);
  function startTest() {
    // Kick off test once audio has loaded.
    video.addEventListener("loadedmetadata", loadedmetadata);
    video.src = "audio.ogg";
  }

  SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
